refactor(movie): 영화 도메인 컴포넌트 단일 파일에서 관리
