+++ src/components/movie.ts
@@ -1,6 +1,21 @@
-import starIconSrc from "../images/star_empty.png";
 import { PopularMovie } from "../types/api";
+import starIconSrc from "../images/star_empty.png";
 import { IMAGE_BASE_URL } from "../utils/constants";
+
+export function createMovieList(movies: PopularMovie[]): HTMLElement {
+  const section = document.createElement("section");
+
+  const ul = document.createElement("ul");
+  ul.className = "thumbnail-list";
+
+  movies.forEach((movie) => {
+    const card = createMovieCard(movie);
+    ul.appendChild(card);
+  });
+
+  section.append(ul);
+  return section;
+}
 
 export function createMovieCard({
   title,
